fixed PagesController tests
tested that all actions except 'show' succeed as admin, return unauthorized as members (non-admins) and redirect to login for visitors (not logged in).  This uncovered a glitch where visitors weren't required to be logged on for pages, but *were* required to have an admin role.

Show action should succeed for admins, members, and visitors.

Fixed Page.pages_menu to return [] instead of nil when it finds an empty set. This fixed a view that was relying on an array.

Fixed index test to check for viewable, editable, and parents arrays, instead of the default pages array.